ingestion-svc/requirements.txt
fastapi
uvicorn
requests
google-cloud-storage
pydantic
pytest
httpx
sqlalchemy
asyncpg
sse-starlette
python-dotenv
aiosqlite